requirements.txt
streamlit>=1.39.0
spotipy>=2.24.0
pandas>=2.2.0
plotly>=5.24.0
python-dotenv>=1.0.0